backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
pydantic==2.7.1
python-dotenv==1.0.1
geoalchemy2==0.15.1
shapely==2.0.4
psycopg2-binary==2.9.9
rio-tiler>=6.0.0
pystac>=1.0.0
pystac-client>=0.7.0
planetary-computer>=1.0.0
httpx>=0.27.0
scikit-learn>=1.4.0
numpy>=1.26.0
geopandas>=1.0.0
pyogrio>=0.10.0
Pillow>=10.0.0
